+++ packages/bbob-parser/src/lexer.js
@@ -11,7 +11,9 @@ import {
   N,
 } from '@bbob/plugin-helper/lib/char';
 
-import { Token, TYPE_ATTR_NAME, TYPE_ATTR_VALUE, TYPE_NEW_LINE, TYPE_SPACE, TYPE_TAG, TYPE_WORD } from './Token';
+import {
+  Token, TYPE_ATTR_NAME, TYPE_ATTR_VALUE, TYPE_NEW_LINE, TYPE_SPACE, TYPE_TAG, TYPE_WORD,
+} from './Token';
 import { createCharGrabber, trimChar, unquote } from './utils';
 
 // for cases <!-- -->
@@ -49,19 +51,22 @@ function createLexer(buffer, options = {}) {
   const tokens = new Array(Math.floor(buffer.length));
   const openTag = options.openTag || OPEN_BRAKET;
   const closeTag = options.closeTag || CLOSE_BRAKET;
+  const escapeTags = options.enableEscapeTags;
 
   const RESERVED_CHARS = [closeTag, openTag, QUOTEMARK, BACKSLASH, SPACE, TAB, EQ, N, EM];
   const NOT_CHAR_TOKENS = [
-    ...(options.enableEscapeTags ? [BACKSLASH] : []),
+    // ...(options.enableEscapeTags ? [BACKSLASH] : []),
     openTag, SPACE, TAB, N,
   ];
   const WHITESPACES = [SPACE, TAB];
   const SPECIAL_CHARS = [EQ, SPACE, TAB];
 
-  const isCharReserved = char => (RESERVED_CHARS.indexOf(char) >= 0);
-  const isWhiteSpace = char => (WHITESPACES.indexOf(char) >= 0);
-  const isCharToken = char => (NOT_CHAR_TOKENS.indexOf(char) === -1);
-  const isSpecialChar = char => (SPECIAL_CHARS.indexOf(char) >= 0);
+  const isCharReserved = (char) => (RESERVED_CHARS.indexOf(char) >= 0);
+  const isWhiteSpace = (char) => (WHITESPACES.indexOf(char) >= 0);
+  const isCharToken = (char) => (NOT_CHAR_TOKENS.indexOf(char) === -1);
+  const isSpecialChar = (char) => (SPECIAL_CHARS.indexOf(char) >= 0);
+  const isEscapableChar = (char) => (char === openTag || char === closeTag || char === BACKSLASH);
+  const isEscapeChar = (char) => char === BACKSLASH;
 
   /**
    * Emits newly created token to subscriber
@@ -158,14 +163,9 @@ function createLexer(buffer, options = {}) {
     } else if (isWhiteSpace(currChar)) {
       const str = bufferGrabber.grabWhile(isWhiteSpace);
       emitToken(createToken(TYPE_SPACE, str, row, col));
-    } else if (options.enableEscapeTags && currChar === BACKSLASH
-               && (nextChar === openTag || nextChar === closeTag)) {
+    } else if (escapeTags && isEscapeChar(currChar) && isEscapableChar(nextChar)) {
       bufferGrabber.skip(); // skip the \ without emitting anything
-      bufferGrabber.skip(); // skip past the [ or ] as well
-      emitToken(createToken(TYPE_WORD, nextChar, row, col));
-    } else if (options.enableEscapeTags && currChar === BACKSLASH && nextChar === BACKSLASH) {
-      bufferGrabber.skip(); // skip the first \ without emitting anything
-      bufferGrabber.skip(); // skip past the second \ and emit it
+      bufferGrabber.skip(); // skip past the [, ] or \ as well
       emitToken(createToken(TYPE_WORD, nextChar, row, col));
     } else if (currChar === openTag) {
       bufferGrabber.skip(); // skip openTag
@@ -177,7 +177,7 @@ function createLexer(buffer, options = {}) {
       if (isCharReserved(nextChar) || hasInvalidChars || bufferGrabber.isLast()) {
         emitToken(createToken(TYPE_WORD, currChar, row, col));
       } else {
-        const str = bufferGrabber.grabWhile(val => val !== closeTag);
+        const str = bufferGrabber.grabWhile((val) => val !== closeTag);
 
         bufferGrabber.skip(); // skip closeTag
         // [myTag   ]
@@ -200,9 +200,19 @@ function createLexer(buffer, options = {}) {
 
       emitToken(createToken(TYPE_WORD, currChar, row, col));
     } else if (isCharToken(currChar)) {
-      const str = bufferGrabber.grabWhile(isCharToken);
+      if (escapeTags && isEscapeChar(currChar) && !isEscapableChar(nextChar)) {
+        bufferGrabber.skip();
+        emitToken(createToken(TYPE_WORD, currChar, row, col));
+      } else {
+        const str = bufferGrabber.grabWhile((char) => {
+          if (escapeTags) {
+            return isCharToken(char) && !isEscapeChar(char);
+          }
+          return isCharToken(char);
+        });
 
-      emitToken(createToken(TYPE_WORD, str, row, col));
+        emitToken(createToken(TYPE_WORD, str, row, col));
+      }
     }
   };