+++ packages/bbob-preset-react/src/index.js
@@ -1,6 +1,6 @@
 import presetHTML5 from '@bbob/preset-html5';
 
-export default presetHTML5.extend(tags => ({
+export default presetHTML5.extend((tags) => ({
   ...tags,
 
   b: (...args) => ({